refactor: rename to inline alert
