+++ src/components/inline-alert/types.ts
@@ -1,13 +1,13 @@
 import { ReactChild } from 'react'
 
-export interface AlertProps {
+export interface InlineAlertProps {
 	children: ReactChild | ReactChild[]
 	type?: 'tip' | 'highlight' | 'note' | 'warning'
 	title?: string
 }
 
-export type AlertData = Record<
-	AlertProps['type'],
+export type InlineAlertData = Record<
+	InlineAlertProps['type'],
 	{
 		title: string
 		icon: JSX.IntrinsicElements['svg']
