+++ src/views/tutorial-view/utils/mdx-components.tsx
@@ -8,7 +8,7 @@ import ImageConfig from 'components/image-config'
 import InteractiveLabCallout from 'components/interactive-lab-callout'
 import {
 	MdxA,
-	MdxAlert,
+	MdxInlineAlert,
 	MdxOrderedList,
 	MdxUnorderedList,
 	MdxListItem,
@@ -45,7 +45,7 @@ const AccordionWrapper = ({ children, collapse, heading }) => {
 //  these components are automatically imported into scope within MDX content
 const MDX_COMPONENTS = {
 	Accordion: AccordionWrapper,
-	Alert: MdxAlert,
+	Alert: MdxInlineAlert,
 	Tabs: MdxTabs,
 	Tab: MdxTab,
 	pre,
